+++ vue/src/Admin/Admin.vue
@@ -525,6 +525,7 @@ export default defineComponent({
       required: true,
     },
     updatedFromPre30: Boolean,
+    requirePasswordConfirmation: Boolean,
   },
   components: {
     AjaxForm,
@@ -584,10 +585,18 @@ export default defineComponent({
     },
     requestSaveLdapConfig() {
       this.pendingSaveTarget = 'config';
+      if (!this.requirePasswordConfirmation) {
+        this.confirmSaveAction();
+        return;
+      }
       this.showPasswordConfirmation = true;
     },
     requestSaveServers() {
       this.pendingSaveTarget = 'servers';
+      if (!this.requirePasswordConfirmation) {
+        this.confirmSaveAction();
+        return;
+      }
       this.showPasswordConfirmation = true;
     },
     confirmSaveAction(passwordConfirmation?: string) {